+++ merakygen/make_powershell_module.py
@@ -69,6 +69,12 @@ def make_function(func_name, func_desc, func_args,
     return function_text
 
 
+def generate_function_docstring(func_docstring):
+    """Based on
+
+    https://docs.microsoft.com/en-us/powershell/module/microsoft.powershell.core/about/about_comment_based_help?view=powershell-6
+    """
+
 def make_classes(api_calls):
     """Add class headers and indent all functions once.